another try to get jna EDSDK access working on Mac
