+++ .gitignore
@@ -10,3 +10,4 @@
 /target
 .settings
 .DS_Store
+hs_err_pid*
