+++ src/main/java/edsdk/api/CanonCommand.java
@@ -41,7 +41,7 @@ import edsdk.utils.CanonUtils;
  */
 public abstract class CanonCommand<T> implements EdsObjectEventHandler {
 
-    public CanonCamera camera;
+    public BaseCanonCamera camera;
     // edsCamera can be reached also through camera.getEdsCamera(), 
     // but it's needed so much that this is a handy shortcut. 
     public EdsCameraRef edsCamera; 
@@ -58,11 +58,11 @@ public abstract class CanonCommand<T> implements EdsObjectEventHandler {
      * The camera is set by the dispatch thread automatically just before run is
      * called.
      * 
-     * @param camera
+     * @param baseCanonCamera
      */
-    public void setCamera( final CanonCamera camera ) {
-        this.camera = camera;
-        this.edsCamera = camera.getEdsCamera(); 
+    public void setCamera( final BaseCanonCamera baseCanonCamera ) {
+        this.camera = baseCanonCamera;
+        this.edsCamera = baseCanonCamera.getEdsCamera(); 
     }
 
     /**
